feat: Refactor GitHub crawler and related services for improved functionality and organization
