+++ src/crawlers/.env.example
@@ -1,31 +1,8 @@
-# GitHub Crawler Configuration
-
-# Required: Azure Storage Account Name
-AZURE_STORAGE_ACCOUNT_NAME=yourstorageaccount
-
-# Required: Azure Storage Account Key
-# Find this in Azure Portal -> Storage Account -> Access Keys
-AZURE_STORAGE_ACCOUNT_KEY=your_storage_account_key_here
-
-# Optional: Container name for storing the markdown files (defaults to 'github-repos')
-AZURE_STORAGE_CONTAINER_NAME=github-repos
-
-# Optional: GitHub Personal Access Token for higher rate limits
-# Without token: 60 requests per hour
-# With token: 5,000 requests per hour
-# Create token at: https://github.com/settings/tokens
-GITHUB_TOKEN=your_github_token_here
-
-
 # Cosmos DB Configuration
-# Required: Cosmos DB Endpoint URL
-# Find this in Azure Portal -> Cosmos DB Account -> Keys
-# Example: https://your-cosmosdb-instance-name.documents.azure.com:443/
 COSMOSDB_ENDPOINT="https://your-cosmosdb-instance-name.documents.azure.com:443/"
-COSMOSDB_KEY=""
-COSMOSDB_DATABASE=""
+COSMOSDB_KEY="your_cosmosdb_key_here"
+COSMOSDB_DATABASE="your_database_name_here"
 
 # Embedding Configuration
 Azure_OPENAI_ENDPOINT="https://your-openai-endpoint.openai.azure.com/"
 Azure_OPENAI_API_KEY="your_openai_api_key_here"
-Azure_OPENAI_DEPLOYMENT_NAME="text-embedding-3-small"
